require Block to have >= 1 Statement
A function with an empty body will confuse the parser.
```
func hello(): void {} %> ParseError
```
This is because `void{}` is a well-formed type: it’s shorthand for `Set.<void>`.
When the parser sees `func hello(): void {}`, it sees a return type
of `void{}` and then expects the rest of the function to follow.
The grammar rule it matches is
```
TypeUnarySymbol ::= TypeUnarySymbol "{" "}";
```

We should keep the tokens `"{"` and `"}"` separate in the grammar,
avoiding adding a new token `"{}"`, for two reasons:

1. Forcing this to be 1 token would mean authors *could not* write `T{ }`
	to mean `Set.<T>`; they would be forced to remove the whitespace.
	This restriction is not in place for `T[ ]`, a.k.a `List.<T>`,
	so it doesn’t make sense to have it for sets.
2. Also, we would have to insert the token as an alternative
	(adding ` | "{}"`) *everywhere* a block appears in the grammar.

Thus the easier way to solve this is to just require blocks to have statements.

To correct the error above, we must insert any statement
(even an empty expression-statement) inside the block delimiters:
```cp
func hello(): void {;} % no parse error
%                   ^ semicolon
```

This change also affects empty blocks in control statements. E.g.,
```
if condition then { result; } else {};  %> ParseError
if condition then { result; } else {;}; % fixed
```
However, it’s not too big of a deal to add a semicolon,
not to mention we can remove the dangling `else` altogether.