perf(tree-sitter-countepoint): flatten operator syntax trees using tree-sitter precedence
note: This new grammar does not strictly adhere to the
grammar spec in syntax.ebnf (Chomsky Type-2, “context-free”),
but is optimized for tree-sitter’s parser generator.

Changing the grammar spec to meet this new implementation
would make it no longer context-free, as it would need to
explicitly support characteristics like precedence and associativity
(whereas those characteristics are implicitly supported by
the the way the CFG productions are written).